refactor(sem): move EventSemantics def to state/mod
