+++ src/event/event_data/receipt.rs
@@ -1,4 +1,4 @@
-use super::EventSemantics;
+use crate::state::EventSemantics;
 use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
